feat: disclamer modal isolation
